ReadingToPercentage Implemented so ADC will read a float between 0 and 1 as a percentage

+++ PiClaw_Client/include/adc.h
@@ -8,7 +8,7 @@ class ADC
         I2Cdev &i2cDev;
         ADS1115 adc1115;
 
-        static int ReadingToPercentage(int reading);
+        static float ReadingToPercentage(float reading);
